Added Basic Typing Controller
This adds a new component that wraps the editor and maintains a queue
of words to be typed, feeding the editor the upcoming words to display
as well.  When all the words have been typed, it fires an event with
the collected results.  In the example file right now I'm just
replacing the editor with a table that displays the results, but I'll
wrap the whole thing in a proper UI shortly.