Fleshing Out the Skeleton
This implements the beginning of the overall app.  There's now a
single component to control the entire page and move through the
possible states (waiting for user to select amount of text to type,
user typing, displaying results).  I've added a basically functional
component that prompts the user for the word count and retrieves the
requested number of words from the server.  There's also a very
rudimentary result display, which will be replaced with something
proper later on down the line.