More Punctuation Spacing
A more thorough look at punctuation spacing, this time taking
quotation characters into account.  It's not perfect because it's
impossible to tell whether a quote mark is opening or closing, but at
least the user should be able to type comfortably because extra spaces
are ignored, even if the formatting is less than optimal.