First Iteration of Color Display
This is the first iteration of the color display, which shows the
words typed color-coded by how far above or below the average speed
(per character) they were typed.  It still needs a lot of work evening
out because of the tendency for very long individual words to screw up
the average, but this should give the general idea.